Run benchmark in a separate workflow
1.0 takes 40+ seconds to run benchmark while 1.1 takes 25+ minutes. It's better to run benchmark as a separate workflow and not by default when doing gradle build.
