+++ src/test/java/org/opensearch/ad/ml/EntityColdStarterTests.java
@@ -29,209 +29,30 @@ package org.opensearch.ad.ml;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import java.io.IOException;
-import java.time.Clock;
-import java.time.Duration;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.AbstractMap.SimpleImmutableEntry;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Queue;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.opensearch.action.ActionListener;
-import org.opensearch.action.get.GetRequest;
-import org.opensearch.action.get.GetResponse;
-import org.opensearch.ad.AbstractADTest;
-import org.opensearch.ad.AnomalyDetectorPlugin;
-import org.opensearch.ad.MemoryTracker;
 import org.opensearch.ad.NodeStateManager;
-import org.opensearch.ad.TestHelpers;
 import org.opensearch.ad.common.exception.AnomalyDetectionException;
-import org.opensearch.ad.dataprocessor.IntegerSensitiveSingleFeatureLinearUniformInterpolator;
-import org.opensearch.ad.dataprocessor.Interpolator;
-import org.opensearch.ad.dataprocessor.LinearUniformInterpolator;
-import org.opensearch.ad.dataprocessor.SingleFeatureLinearUniformInterpolator;
-import org.opensearch.ad.feature.FeatureManager;
-import org.opensearch.ad.feature.SearchFeatureDao;
 import org.opensearch.ad.ml.ModelManager.ModelType;
-import org.opensearch.ad.model.AnomalyDetector;
-import org.opensearch.ad.model.IntervalTimeConfiguration;
-import org.opensearch.ad.settings.AnomalyDetectorSettings;
-import org.opensearch.ad.util.ClientUtil;
-import org.opensearch.ad.util.LabelledAnomalyGenerator;
-import org.opensearch.ad.util.MultiDimDataWithTime;
-import org.opensearch.client.Client;
-import org.opensearch.common.settings.Settings;
 import org.opensearch.common.util.concurrent.OpenSearchRejectedExecutionException;
-import org.opensearch.threadpool.ThreadPool;
 
 import test.org.opensearch.ad.util.MLUtil;
 
 import com.amazon.randomcutforest.RandomCutForest;
-import com.amazon.randomcutforest.serialize.RandomCutForestSerDe;
-import com.google.common.collect.ImmutableList;
-import com.google.gson.Gson;
 
-public class EntityColdStarterTests extends AbstractADTest {
-    int numMinSamples;
-    String modelId;
-    String entityName;
-    String detectorId;
-    ModelState<EntityModel> modelState;
-    Clock clock;
-    float priority;
-    EntityColdStarter entityColdStarter;
-    NodeStateManager stateManager;
-    SearchFeatureDao searchFeatureDao;
-    Interpolator interpolator;
-    CheckpointDao checkpoint;
-    FeatureManager featureManager;
-    Settings settings;
-    ThreadPool threadPool;
-    Runnable releaseSemaphore;
-    AnomalyDetector detector;
-    long rcfSeed;
-    ClientUtil clientUtil;
-    ModelManager modelManager;
-    Client client;
-    ModelPartitioner modelPartitioner;
-
-    @SuppressWarnings("unchecked")
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
-        numMinSamples = AnomalyDetectorSettings.NUM_MIN_SAMPLES;
-
-        clock = mock(Clock.class);
-        when(clock.instant()).thenReturn(Instant.now());
-
-        threadPool = mock(ThreadPool.class);
-        setUpADThreadPool(threadPool);
-
-        settings = Settings.EMPTY;
-
-        client = mock(Client.class);
-        clientUtil = mock(ClientUtil.class);
-
-        AnomalyDetector detector = TestHelpers
-            .randomAnomalyDetectorWithInterval(new IntervalTimeConfiguration(1, ChronoUnit.MINUTES), true, true);
-        doAnswer(invocation -> {
-            ActionListener<GetResponse> listener = invocation.getArgument(2);
-            listener.onResponse(TestHelpers.createGetResponse(detector, detectorId, AnomalyDetector.ANOMALY_DETECTORS_INDEX));
-            return null;
-        }).when(clientUtil).asyncRequest(any(GetRequest.class), any(), any(ActionListener.class));
-
-        modelPartitioner = mock(ModelPartitioner.class);
-        stateManager = new NodeStateManager(
-            client,
-            xContentRegistry(),
-            settings,
-            clientUtil,
-            clock,
-            AnomalyDetectorSettings.HOURLY_MAINTENANCE,
-            modelPartitioner
-        );
-
-        SingleFeatureLinearUniformInterpolator singleFeatureLinearUniformInterpolator =
-            new IntegerSensitiveSingleFeatureLinearUniformInterpolator();
-        interpolator = new LinearUniformInterpolator(singleFeatureLinearUniformInterpolator);
-
-        searchFeatureDao = mock(SearchFeatureDao.class);
-        checkpoint = mock(CheckpointDao.class);
-
-        featureManager = new FeatureManager(
-            searchFeatureDao,
-            interpolator,
-            clock,
-            AnomalyDetectorSettings.MAX_TRAIN_SAMPLE,
-            AnomalyDetectorSettings.MAX_SAMPLE_STRIDE,
-            AnomalyDetectorSettings.TRAIN_SAMPLE_TIME_RANGE_IN_HOURS,
-            AnomalyDetectorSettings.MIN_TRAIN_SAMPLES,
-            AnomalyDetectorSettings.MAX_SHINGLE_PROPORTION_MISSING,
-            AnomalyDetectorSettings.MAX_IMPUTATION_NEIGHBOR_DISTANCE,
-            AnomalyDetectorSettings.PREVIEW_SAMPLE_RATE,
-            AnomalyDetectorSettings.MAX_PREVIEW_SAMPLES,
-            AnomalyDetectorSettings.HOURLY_MAINTENANCE,
-            threadPool,
-            AnomalyDetectorPlugin.AD_THREAD_POOL_NAME
-        );
-
-        entityColdStarter = new EntityColdStarter(
-            clock,
-            threadPool,
-            stateManager,
-            AnomalyDetectorSettings.NUM_SAMPLES_PER_TREE,
-            AnomalyDetectorSettings.MULTI_ENTITY_NUM_TREES,
-            AnomalyDetectorSettings.TIME_DECAY,
-            numMinSamples,
-            AnomalyDetectorSettings.MAX_SAMPLE_STRIDE,
-            AnomalyDetectorSettings.MAX_TRAIN_SAMPLE,
-            interpolator,
-            searchFeatureDao,
-            AnomalyDetectorSettings.DEFAULT_MULTI_ENTITY_SHINGLE,
-            AnomalyDetectorSettings.THRESHOLD_MIN_PVALUE,
-            AnomalyDetectorSettings.THRESHOLD_MAX_RANK_ERROR,
-            AnomalyDetectorSettings.THRESHOLD_MAX_SCORE,
-            AnomalyDetectorSettings.THRESHOLD_NUM_LOGNORMAL_QUANTILES,
-            AnomalyDetectorSettings.THRESHOLD_DOWNSAMPLES,
-            AnomalyDetectorSettings.THRESHOLD_MAX_SAMPLES,
-            featureManager,
-            AnomalyDetectorSettings.HOURLY_MAINTENANCE,
-            AnomalyDetectorSettings.MAX_SMALL_STATES,
-            checkpoint,
-            settings
-        );
-
-        detectorId = "123";
-        modelId = "123_entity_abc";
-        entityName = "abc";
-        priority = 0.3f;
-
-        RandomCutForestSerDe rcfSerde = mock(RandomCutForestSerDe.class);
-        MemoryTracker memoryTracker = mock(MemoryTracker.class);
-
-        modelManager = new ModelManager(
-            rcfSerde,
-            checkpoint,
-            new Gson(),
-            clock,
-            AnomalyDetectorSettings.NUM_TREES,
-            AnomalyDetectorSettings.NUM_SAMPLES_PER_TREE,
-            AnomalyDetectorSettings.TIME_DECAY,
-            numMinSamples,
-            AnomalyDetectorSettings.THRESHOLD_MIN_PVALUE,
-            AnomalyDetectorSettings.THRESHOLD_MAX_RANK_ERROR,
-            AnomalyDetectorSettings.THRESHOLD_MAX_SCORE,
-            AnomalyDetectorSettings.THRESHOLD_NUM_LOGNORMAL_QUANTILES,
-            AnomalyDetectorSettings.THRESHOLD_DOWNSAMPLES,
-            AnomalyDetectorSettings.THRESHOLD_MAX_SAMPLES,
-            HybridThresholdingModel.class,
-            500,
-            Duration.ofHours(1),
-            Duration.ofHours(1),
-            entityColdStarter,
-            modelPartitioner,
-            featureManager,
-            memoryTracker
-        );
-    }
-
+public class EntityColdStarterTests extends AbstractCosineDataTest {
     // train using samples directly
     public void testTrainUsingSamples() {
         Queue<double[]> samples = MLUtil.createQueueSamples(numMinSamples);
@@ -287,16 +108,6 @@ public class EntityColdStarterTests extends AbstractADTest {
 
         assertTrue(forest == null);
         assertTrue(model.getThreshold() == null);
-    }
-
-    private void waitForColdStartFinish() throws InterruptedException {
-        int maxWaitTimes = 20;
-        int i = 0;
-        while (stateManager.isColdStartRunning(detectorId) && i < maxWaitTimes) {
-            // wait for 500 milliseconds
-            Thread.sleep(500L);
-            i++;
-        }
     }
 
     // cold start running, return immediately
@@ -505,251 +316,5 @@ public class EntityColdStarterTests extends AbstractADTest {
         assertTrue(model.getThreshold() == null);
         // the min-max range is too small and thus no data range can be found
         assertEquals("real sample size is " + model.getSamples().size(), 1, model.getSamples().size());
-    }
-
-    public int searchInsert(long[] timestamps, long target) {
-        int pivot, left = 0, right = timestamps.length - 1;
-        while (left <= right) {
-            pivot = left + (right - left) / 2;
-            if (timestamps[pivot] == target)
-                return pivot;
-            if (target < timestamps[pivot])
-                right = pivot - 1;
-            else
-                left = pivot + 1;
-        }
-        return left;
-    }
-
-    /**
-     * A template to perform precision/recall test.
-     *
-     * @param detectorIntervalMins Detector interval
-     * @param precisionThreshold precision threshold
-     * @param recallThreshold recall threshold
-     * @param baseDimension the number of dimensions
-     * @param anomalyIndependent whether anomalies in each dimension is generated independently
-     * @throws Exception when failing to create anomaly detector or creating training data
-     */
-    @SuppressWarnings("unchecked")
-    private void averageAccuracyTemplate(
-        int detectorIntervalMins,
-        float precisionThreshold,
-        float recallThreshold,
-        int baseDimension,
-        boolean anomalyIndependent
-    ) throws Exception {
-        int dataSize = 20 * AnomalyDetectorSettings.NUM_SAMPLES_PER_TREE;
-        int trainTestSplit = 300;
-        // detector interval
-        int interval = detectorIntervalMins;
-        int delta = 60000 * interval;
-
-        int numberOfTrials = 10;
-        double prec = 0;
-        double recall = 0;
-        double totalPrec = 0;
-        double totalRecall = 0;
-
-        // training data ranges from timestamps[0] ~ timestamps[trainTestSplit-1]
-        // set up detector
-        detector = TestHelpers.AnomalyDetectorBuilder
-            .newInstance()
-            .setDetectionInterval(new IntervalTimeConfiguration(interval, ChronoUnit.MINUTES))
-            .setCategoryFields(ImmutableList.of(randomAlphaOfLength(5)))
-            .setShingleSize(AnomalyDetectorSettings.DEFAULT_MULTI_ENTITY_SHINGLE)
-            .build();
-
-        doAnswer(invocation -> {
-            ActionListener<GetResponse> listener = invocation.getArgument(2);
-
-            listener.onResponse(TestHelpers.createGetResponse(detector, detector.getDetectorId(), AnomalyDetector.ANOMALY_DETECTORS_INDEX));
-            return null;
-        }).when(clientUtil).asyncRequest(any(GetRequest.class), any(), any(ActionListener.class));
-
-        for (int z = 1; z <= numberOfTrials; z++) {
-            long seed = z;
-            LOG.info("seed = " + seed);
-
-            searchFeatureDao = mock(SearchFeatureDao.class);
-
-            featureManager = new FeatureManager(
-                searchFeatureDao,
-                interpolator,
-                clock,
-                AnomalyDetectorSettings.MAX_TRAIN_SAMPLE,
-                AnomalyDetectorSettings.MAX_SAMPLE_STRIDE,
-                AnomalyDetectorSettings.TRAIN_SAMPLE_TIME_RANGE_IN_HOURS,
-                AnomalyDetectorSettings.MIN_TRAIN_SAMPLES,
-                AnomalyDetectorSettings.MAX_SHINGLE_PROPORTION_MISSING,
-                AnomalyDetectorSettings.MAX_IMPUTATION_NEIGHBOR_DISTANCE,
-                AnomalyDetectorSettings.PREVIEW_SAMPLE_RATE,
-                AnomalyDetectorSettings.MAX_PREVIEW_SAMPLES,
-                AnomalyDetectorSettings.HOURLY_MAINTENANCE,
-                threadPool,
-                AnomalyDetectorPlugin.AD_THREAD_POOL_NAME
-            );
-
-            entityColdStarter = new EntityColdStarter(
-                clock,
-                threadPool,
-                stateManager,
-                AnomalyDetectorSettings.NUM_SAMPLES_PER_TREE,
-                AnomalyDetectorSettings.MULTI_ENTITY_NUM_TREES,
-                AnomalyDetectorSettings.TIME_DECAY,
-                numMinSamples,
-                AnomalyDetectorSettings.MAX_SAMPLE_STRIDE,
-                AnomalyDetectorSettings.MAX_TRAIN_SAMPLE,
-                interpolator,
-                searchFeatureDao,
-                AnomalyDetectorSettings.DEFAULT_MULTI_ENTITY_SHINGLE,
-                AnomalyDetectorSettings.THRESHOLD_MIN_PVALUE,
-                AnomalyDetectorSettings.THRESHOLD_MAX_RANK_ERROR,
-                AnomalyDetectorSettings.THRESHOLD_MAX_SCORE,
-                AnomalyDetectorSettings.THRESHOLD_NUM_LOGNORMAL_QUANTILES,
-                AnomalyDetectorSettings.THRESHOLD_DOWNSAMPLES,
-                AnomalyDetectorSettings.THRESHOLD_MAX_SAMPLES,
-                featureManager,
-                AnomalyDetectorSettings.HOURLY_MAINTENANCE,
-                AnomalyDetectorSettings.MAX_SMALL_STATES,
-                checkpoint,
-                settings,
-                seed
-            );
-
-            // create labelled data
-            MultiDimDataWithTime dataWithKeys = LabelledAnomalyGenerator
-                .getMultiDimData(
-                    dataSize + detector.getShingleSize() - 1,
-                    50,
-                    100,
-                    5,
-                    seed,
-                    baseDimension,
-                    false,
-                    trainTestSplit,
-                    delta,
-                    anomalyIndependent
-                );
-
-            long[] timestamps = dataWithKeys.timestampsMs;
-            double[][] data = dataWithKeys.data;
-            when(clock.millis()).thenReturn(timestamps[trainTestSplit - 1]);
-
-            doAnswer(invocation -> {
-                ActionListener<Entry<Optional<Long>, Optional<Long>>> listener = invocation.getArgument(2);
-                listener.onResponse(new SimpleImmutableEntry<>(Optional.of(timestamps[0]), Optional.of(timestamps[trainTestSplit - 1])));
-                return null;
-            }).when(searchFeatureDao).getEntityMinMaxDataTime(any(), any(), any());
-
-            doAnswer(invocation -> {
-                List<Entry<Long, Long>> ranges = invocation.getArgument(1);
-                List<Optional<double[]>> coldStartSamples = new ArrayList<>();
-
-                Collections.sort(ranges, new Comparator<Entry<Long, Long>>() {
-                    @Override
-                    public int compare(Entry<Long, Long> p1, Entry<Long, Long> p2) {
-                        return Long.compare(p1.getKey(), p2.getKey());
-                    }
-                });
-                for (int j = 0; j < ranges.size(); j++) {
-                    Entry<Long, Long> range = ranges.get(j);
-                    Long start = range.getKey();
-                    int valueIndex = searchInsert(timestamps, start);
-                    coldStartSamples.add(Optional.of(data[valueIndex]));
-                }
-
-                ActionListener<List<Optional<double[]>>> listener = invocation.getArgument(4);
-                listener.onResponse(coldStartSamples);
-                return null;
-            }).when(searchFeatureDao).getColdStartSamplesForPeriods(any(), any(), any(), anyBoolean(), any());
-
-            String entityNamez = entityName + z;
-            String modelId = detector.getDetectorId() + "_entity_" + entityNamez;
-            EntityModel model = new EntityModel(modelId, new ArrayDeque<>(), null, null);
-            ModelState<EntityModel> modelState = new ModelState<>(
-                model,
-                modelId,
-                detector.getDetectorId(),
-                ModelType.ENTITY.getName(),
-                clock,
-                priority
-            );
-
-            entityColdStarter.trainModel(new ArrayDeque<>(), modelId, entityNamez, detector.getDetectorId(), modelState);
-
-            waitForColdStartFinish();
-            assertTrue(model.getRcf() != null);
-
-            int tp = 0;
-            int fp = 0;
-            int fn = 0;
-            long[] changeTimestamps = dataWithKeys.changeTimeStampsMs;
-
-            for (int j = trainTestSplit; j < data.length; j++) {
-                ThresholdingResult result = modelManager
-                    .getAnomalyResultForEntity(detector.getDetectorId(), data[j], entityNamez, modelState, modelId);
-                if (result.getGrade() > 0) {
-                    if (changeTimestamps[j] == 0) {
-                        fp++;
-                    } else {
-                        tp++;
-                    }
-                } else {
-                    if (changeTimestamps[j] != 0) {
-                        fn++;
-                    }
-                    // else ok
-                }
-            }
-
-            if (tp + fp == 0) {
-                prec = 1;
-            } else {
-                prec = tp * 1.0 / (tp + fp);
-            }
-
-            if (tp + fn == 0) {
-                recall = 1;
-            } else {
-                recall = tp * 1.0 / (tp + fn);
-            }
-
-            totalPrec += prec;
-            totalRecall += recall;
-            modelState = null;
-            dataWithKeys = null;
-            reset(searchFeatureDao);
-            searchFeatureDao = null;
-        }
-
-        double avgPrec = totalPrec / numberOfTrials;
-        double avgRecall = totalRecall / numberOfTrials;
-        LOG.info("{} features, Interval {}, Precision: {}, recall: {}", baseDimension, detectorIntervalMins, avgPrec, avgRecall);
-        assertTrue("average precision is " + avgPrec, avgPrec >= precisionThreshold);
-        assertTrue("average recall is " + avgRecall, avgRecall >= recallThreshold);
-    }
-
-    public void testAverageAccuracy() throws Exception {
-        for (final boolean anomalyIndependent : new boolean[] { false, true }) {
-            LOG.info("Anomalies independence: " + anomalyIndependent);
-            // 10 minute interval, 4 features
-            averageAccuracyTemplate(10, 0.8f, 0.1f, 4, anomalyIndependent);
-
-            // 10 minute interval, 2 features
-            averageAccuracyTemplate(10, 0.7f, 0.2f, 2, anomalyIndependent);
-
-            // 10 minute interval, 1 features
-            averageAccuracyTemplate(10, 0.5f, 0.2f, 1, anomalyIndependent);
-
-            // 5 minute interval, 4 features
-            averageAccuracyTemplate(5, 0.3f, 0.1f, 4, anomalyIndependent);
-
-            // 5 minute interval, 2 features
-            averageAccuracyTemplate(5, 0.7f, 0.2f, 2, anomalyIndependent);
-
-            // 5 minute interval, 1 features
-            averageAccuracyTemplate(5, 0.5f, 0.2f, 1, anomalyIndependent);
-        }
     }
 }
